#603 - fix ParserExtension
