+++ coffee-module/coffee-module-mp/coffee-module-mp-restclient/src/main/resources/META-INF/services/jakarta.enterprise.inject.spi.Extension
@@ -1,0 +1,1 @@
+hu.icellmobilsoft.coffee.module.mp.restclient.exception.FaultTypeParserExtension
